LayoutTests/touchadjustment/search-cancel.html

<html>
<head>
<style>
    input[type='search'] {
      -webkit-appearance: none;
      -webkit-box-sizing: border-box;
      border: 1px solid #999;
      padding: 5px;
    }

    input[type='search']::-webkit-search-cancel-button {
      -webkit-appearance: none;
      -webkit-box-sizing: border-box;
      background: #999;
      height: 10px;
      margin: 0;
      padding: 0;
      width: 10px;
    }
</style>
<script src="../fast/js/resources/js-test-pre.js"></script>

</head>

<body onload="runTests()">

<div>
  <input id="target" type="search" value="abcd"/>
</div>

<p id='description'></p>
<div id='console'></div>

<script>
    var element;
    var adjustedNode;
    var adjustedPoint;
    var inputBounds;
    var cancelBounds;
    var textContentBounds;
    var touchBounds;
    var targetBounds;
    var targetContainsPoint;
    var touchContainsPoint;

    function findAbsolutePosition(node) {
        var pos = {left: 0, top: 0};
        do {
            pos.left += node.offsetLeft;
            pos.top += node.offsetTop;
        } while (node = node.offsetParent);
        return pos;
    }

    function runTouchTests() {
        document.addEventListener('click', function() {}, false);

        element = document.getElementById('target');
        var pos = findAbsolutePosition(element);

        var x = pos.left;
        var y = pos.top;
        var width = element.clientWidth;
        var height = element.clientHeight;

        inputBounds = {
            x: x,
            y: y,
            width: width,
            height: height
        };

        var touchRadius = 20;
        var offset = touchRadius / 2;

        // Center of cancel button.
        var midX = x + width - 11;
        var midY = y + height / 2;

        cancelBounds = {
            x: midX - 5,
            y: midY - 5,
            width: 10,
            height: 10
        };

        textContentBounds = {
           x: inputBounds.x,
           y: inputBounds.y,
           width: inputBounds.width - 16,
           height: inputBounds.height
        };

        debug('\nTouch centered on cancel button but overlapping neighboring candidate.');
        testTouch(midX, midY, touchRadius, cancelBounds);

        debug('\nOff-center touches enclosing cancel button.');
        testTouch(midX - offset, midY, touchRadius, cancelBounds);
        testTouch(midX - offset, midY + offset, touchRadius, cancelBounds);
        testTouch(midX - offset, midY - offset, touchRadius, cancelBounds);

        debug('\nTouches clipping left edge of cancel-button.');
        testTouch(midX - touchRadius - 2, midY, touchRadius, textContentBounds);
        testTouch(midX - touchRadius - 2, midY - offset, touchRadius, textContentBounds);
        testTouch(midX - touchRadius - 2, midY + offset, touchRadius, textContentBounds);
    }

    function testTouch(touchX, touchY, radius, bounds) {

        targetBounds = bounds;

        touchBounds = {
            x: touchX - radius,
            y: touchY - radius,
            width: 2 * radius,
            height: 2 * radius
        };
        adjustedPoint = internals.touchPositionAdjustedToBestClickableNode(touchBounds.x, touchBounds.y, touchBounds.width, touchBounds.height, document);
        targetContainsPoint = containsPoint(targetBounds, adjustedPoint);
        touchContainsPoint = containsPoint(touchBounds, adjustedPoint);
        shouldBeTrue('targetContainsPoint');
        shouldBeTrue('touchContainsPoint');
    }

    function containsPoint(bounds, point) {
        return bounds.x <= point.x && bounds.y <= point.y && bounds.x + bounds.width >= point.x && bounds.y + bounds.height >= point.y;
    }

    function runTests()
    {
        if (window.testRunner && window.internals && internals.touchNodeAdjustedToBestClickableNode) {
            description('Test touch adjustment on a search field with a visible cancel button.');
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
            runTouchTests();
            isSuccessfullyParsed();
            testRunner.notifyDone();
        }
    }
</script>
</body>
</html>
